+++ app/src/js/components/Comments/comments.js
@@ -35,6 +35,13 @@ class Comment extends React.Component {
     return words.join(' ');
   }
 
+  formatComments () {
+    if (document.querySelectorAll('textarea#comment')[0].value !== '') {
+      document.querySelectorAll('textarea#comment')[0].placeholder = 'Enter a comment';
+      document.querySelectorAll('textarea#comment')[0].classList.remove('required');
+    }
+  }
+
   hasStepData () {
     if (typeof this.props.requests !== 'undefined' &&
     typeof this.props.requests.detail.data !== 'undefined' &&
@@ -93,12 +100,11 @@ class Comment extends React.Component {
                 onSubmit={(e) => { e.preventDefault(); this.reply(requestName, conversationId, stepName); }}>
                 <span id='previously-saved' style={{ padding: '0.3em 2em 0.4em 0.7em' }}></span>
                 <textarea placeholder='Enter a comment'
-                  className='no-resize'
                   ref={this.state.textRef}
                   id='comment'
                   aria-label="Enter a comment"
                   title="Enter a comment"
-                  onChange={(e) => { e.preventDefault(); document.querySelectorAll('button.button--reply')[0].classList.remove('hidden'); }}
+                  onChange={(e) => { e.preventDefault(); this.formatComments(); document.querySelectorAll('button.button--reply')[0].classList.remove('hidden'); }}
                   ></textarea>
                 <div style={{ minHeight: '40px' }}>
                   <button type='submit'