+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 <!-- Unreleased changes can be added here. -->
 - Added functionality allowing users to see forms throughout the entire workflow.
+- Made textareas expandable
+- Added 'New Request' button to 'Home' per info team request.
 
 ## 1.0.12 - 2023-05-12
 - Added comment component for of GES DISC workflow requests